av1/common/x86/warp_plane_avx2: fix int sanitizer warnings
implicit conversion from type 'int' of value 33792 (32-bit, signed) to
type 'short' changed the value to -31744 (16-bit, signed)

BUG=aomedia:2357

+++ av1/common/x86/warp_plane_avx2.c
@@ -418,8 +418,8 @@ static INLINE void unpack_weights_and_set_round_const_avx2(
 
   const int w0 = conv_params->fwd_offset;
   const int w1 = conv_params->bck_offset;
-  const __m256i wt0 = _mm256_set1_epi16(w0);
-  const __m256i wt1 = _mm256_set1_epi16(w1);
+  const __m256i wt0 = _mm256_set1_epi16((short)w0);
+  const __m256i wt1 = _mm256_set1_epi16((short)w1);
   *wt = _mm256_unpacklo_epi16(wt0, wt1);
 }
